+++ sqflite/android/src/main/java/com/tekartik/sqflite/SqflitePlugin.java
@@ -27,6 +27,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import io.flutter.embedding.engine.plugins.FlutterPlugin;
+import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
@@ -40,6 +42,7 @@ import static com.tekartik.sqflite.Constant.METHOD_BATCH;
 import static com.tekartik.sqflite.Constant.METHOD_CLOSE_DATABASE;
 import static com.tekartik.sqflite.Constant.METHOD_DEBUG;
 import static com.tekartik.sqflite.Constant.METHOD_DEBUG_MODE;
+import static com.tekartik.sqflite.Constant.METHOD_DELETE_DATABASE;
 import static com.tekartik.sqflite.Constant.METHOD_EXECUTE;
 import static com.tekartik.sqflite.Constant.METHOD_GET_DATABASES_PATH;
 import static com.tekartik.sqflite.Constant.METHOD_GET_PLATFORM_VERSION;
@@ -66,7 +69,8 @@ import static com.tekartik.sqflite.Constant.TAG;
 /**
  * SqflitePlugin Android implementation
  */
-public class SqflitePlugin implements MethodCallHandler {
+public class SqflitePlugin implements FlutterPlugin, MethodCallHandler {
+
 
     static final Map<String, Integer> _singleInstancesByPath = new HashMap<>();
     static private boolean QUERY_AS_MAP_LIST = false; // set by options
@@ -77,26 +81,50 @@ public class SqflitePlugin implements MethodCallHandler {
     static private final Object openCloseLocker = new Object();
     // local cache
     static String databasesPath;
-    static private Context context;
+    private Context context;
     static private int databaseId = 0; // incremental database id
     // Database thread execution
     static private HandlerThread handlerThread;
     static private Handler handler;
-
+    private MethodChannel methodChannel;
     @SuppressLint("UseSparseArrays")
     static final Map<Integer, Database> databaseMap = new HashMap<>();
 
-    SqflitePlugin(Context context) {
-        this.context = context;
+    // Needed public constructor
+    public SqflitePlugin() {
+
+    }
+
+    // Testing only
+    public SqflitePlugin(Context context) {
+        this.context = context.getApplicationContext();
     }
 
     //
     // Plugin registration.
     //
     public static void registerWith(Registrar registrar) {
-        SQLiteDatabase.loadLibs(registrar.context());
-        final MethodChannel channel = new MethodChannel(registrar.messenger(), "com.tekartik.sqflite");
-        channel.setMethodCallHandler(new SqflitePlugin(registrar.context()));
+        SqflitePlugin sqflitePlugin = new SqflitePlugin();
+        sqflitePlugin.onAttachedToEngine(registrar.context(), registrar.messenger());
+    }
+
+    @Override
+    public void onAttachedToEngine(FlutterPluginBinding binding) {
+        onAttachedToEngine(binding.getApplicationContext(), binding.getBinaryMessenger());
+    }
+
+    private void onAttachedToEngine(Context applicationContext, BinaryMessenger messenger) {
+        this.context = applicationContext;
+        SQLiteDatabase.loadLibs(applicationContext);
+        methodChannel = new MethodChannel(messenger, Constant.PLUGIN_KEY);
+        methodChannel.setMethodCallHandler(this);
+    }
+
+    @Override
+    public void onDetachedFromEngine(FlutterPluginBinding binding) {
+        context = null;
+        methodChannel.setMethodCallHandler(null);
+        methodChannel = null;
     }
 
     private static Object cursorValue(Cursor cursor, int index) {
@@ -832,30 +860,97 @@ public class SqflitePlugin implements MethodCallHandler {
             @Override
             public void run() {
                 synchronized (openCloseLocker) {
-
-                    try {
-                        database.close();
-                    } catch (Exception e) {
-                        Log.e(TAG, "error " + e + " while closing database " + databaseId);
-                    }
-
-                    synchronized (databaseMapLocker) {
-
-                        if (databaseMap.isEmpty() && handler != null) {
-                            if (LogLevel.hasSqlLevel(database.logLevel)) {
-                                Log.d(TAG, database.getThreadLogPrefix() + "stopping thread" + handlerThread);
-                            }
-                            handlerThread.quit();
-                            handlerThread = null;
-                            handler = null;
-                        }
-                    }
+                    closeDatabase(database);
                 }
 
                 bgResult.success(null);
             }
         });
 
+    }
+
+    //
+    // Sqflite.open
+    //
+    private void onDeleteDatabaseCall(final MethodCall call, Result result) {
+        final String path = call.argument(PARAM_PATH);
+        Database foundOpenedDatabase = null;
+        // Look for in memory instance
+        synchronized (databaseMapLocker) {
+            if (LogLevel.hasVerboseLevel(logLevel)) {
+                Log.d(Constant.TAG, "Look for " + path + " in " + _singleInstancesByPath.keySet());
+            }
+            Integer databaseId = _singleInstancesByPath.get(path);
+            if (databaseId != null) {
+                Database database = databaseMap.get(databaseId);
+                if (database != null) {
+                    if (database.sqliteDatabase.isOpen()) {
+                        if (LogLevel.hasVerboseLevel(logLevel)) {
+                            Log.d(Constant.TAG, database.getThreadLogPrefix() + "found single instance " + (database.inTransaction ? "(in transaction) " : "") + databaseId + " " + path);
+                        }
+                        foundOpenedDatabase = database;
+
+                        // Remove from map right away
+                        databaseMap.remove(databaseId);
+                        _singleInstancesByPath.remove(path);
+                    }
+                }
+            }
+        }
+        final Database openedDatabase = foundOpenedDatabase;
+
+        final BgResult bgResult = new BgResult(result);
+        final Runnable deleteRunnable = new Runnable() {
+            @Override
+            public void run() {
+                synchronized (openCloseLocker) {
+
+                    if (openedDatabase != null) {
+                        closeDatabase(openedDatabase);
+                    }
+                    try {
+                        if (LogLevel.hasVerboseLevel(logLevel)) {
+                            Log.d(Constant.TAG, "delete database " + path);
+                        }
+                        Database.deleteDatabase(path);
+                    } catch (Exception e) {
+                        Log.e(TAG, "error " + e + " while closing database " + databaseId);
+                    }
+                }
+                bgResult.success(null);
+            }
+        };
+
+        // handler might not exist yet
+        if (handler != null) {
+            handler.post(deleteRunnable);
+        } else {
+            // Otherwise run in the UI thread
+            deleteRunnable.run();
+        }
+
+    }
+
+    private void closeDatabase(Database database) {
+        try {
+            if (LogLevel.hasSqlLevel(database.logLevel)) {
+                Log.d(TAG, database.getThreadLogPrefix() + "closing database " + handlerThread);
+            }
+            database.close();
+        } catch (Exception e) {
+            Log.e(TAG, "error " + e + " while closing database " + databaseId);
+        }
+        synchronized (databaseMapLocker) {
+
+            if (databaseMap.isEmpty() && handler != null) {
+                if (LogLevel.hasSqlLevel(database.logLevel)) {
+                    Log.d(TAG, database.getThreadLogPrefix() + "stopping thread" + handlerThread);
+                }
+                handlerThread.quit();
+                handlerThread = null;
+                handler = null;
+            }
+        }
     }
 
     @Override
@@ -899,7 +994,11 @@ public class SqflitePlugin implements MethodCallHandler {
                 break;
             }
             case METHOD_GET_DATABASES_PATH: {
-                onGetDatabasesPath(call, result);
+                onGetDatabasesPathCall(call, result);
+                break;
+            }
+            case METHOD_DELETE_DATABASE: {
+                onDeleteDatabaseCall(call, result);
                 break;
             }
             case METHOD_DEBUG: {
@@ -935,7 +1034,7 @@ public class SqflitePlugin implements MethodCallHandler {
 
     //private static class Database
 
-    void onGetDatabasesPath(final MethodCall call, Result result) {
+    void onGetDatabasesPathCall(final MethodCall call, Result result) {
         if (databasesPath == null) {
             String dummyDatabaseName = "tekartik_sqflite.db";
             File file = context.getDatabasePath(dummyDatabaseName);
@@ -943,7 +1042,6 @@ public class SqflitePlugin implements MethodCallHandler {
         }
         result.success(databasesPath);
     }
-
 
 
     private class BgResult implements Result {