+++ sqflite/lib/src/batch.dart
@@ -1,9 +1,9 @@
 import 'package:sqflite/sqlite_api.dart';
 import 'package:sqflite/src/constant.dart';
 import 'package:sqflite/src/database.dart';
-import 'package:sqflite/src/sqflite_impl.dart';
 import 'package:sqflite/src/sql_builder.dart';
 import 'package:sqflite/src/transaction.dart';
+import 'package:sqflite/src/utils.dart';
 
 /// Batch implementation
 abstract class SqfliteBatch implements Batch {
@@ -143,8 +143,8 @@ class SqfliteTransactionBatch extends SqfliteBatch {
   Future<List<dynamic>> commit(
       {bool exclusive, bool noResult, bool continueOnError}) {
     if (exclusive != null) {
-      throw ArgumentError.value(exclusive, "exclusive",
-          "must not be set when commiting a batch in a transaction");
+      throw ArgumentError.value(exclusive, 'exclusive',
+          'must not be set when commiting a batch in a transaction');
     }
     return transaction.database.txnApplyBatch(transaction, this,
         noResult: noResult, continueOnError: continueOnError);