+++ sqflite/lib/src/exception_impl.dart
@@ -1,9 +1,9 @@
-import 'package:flutter/services.dart';
+import 'package:sqflite/src/services_impl.dart';
 import 'package:sqflite/src/constant.dart';
 import 'package:sqflite/src/exception.dart';
 
 /// Wrap any exception to a [DatabastException]
-Future<T> wrapDatabaseException<T>(Future<T> action()) async {
+Future<T> wrapDatabaseException<T>(Future<T> Function() action) async {
   try {
     final T result = await action();
     return result;