+++ sqflite/lib/src/factory.dart
@@ -2,15 +2,14 @@ import 'dart:async';
 
 import 'package:sqflite/sqlite_api.dart';
 import 'package:sqflite/src/database.dart';
+import 'package:sqflite/src/mixin/factory.dart';
 import 'package:synchronized/synchronized.dart';
 
 /// Internal database factory interface.
-abstract class SqfliteDatabaseFactory implements DatabaseFactory {
-  /// Invoke a native method.
-  Future<T> invokeMethod<T>(String method, [dynamic arguments]);
-
+abstract class SqfliteDatabaseFactory
+    implements DatabaseFactory, SqfliteInvokeHandler {
   /// Wrap any exception to a [DatabaseException].
-  Future<T> wrapDatabaseException<T>(Future<T> action());
+  Future<T> wrapDatabaseException<T>(Future<T> Function() action);
   // To override
   // This also should wrap exception
   //Future<T> safeInvokeMethod<T>(String method, [dynamic arguments]);
@@ -38,7 +37,4 @@ abstract class SqfliteDatabaseFactory implements DatabaseFactory {
 
   @override
   Future<bool> databaseExists(String path);
-
-  /// Create the parent directory of a database.
-  Future<void> createParentDirectory(String path);
 }