+++ sqflite/test/sqlite_api_test.dart
@@ -2,9 +2,9 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:sqflite/sqlite_api.dart';
 
 void main() {
-  group("sqlite_api", () {
+  group('sqlite_api', () {
     // Check that public api are exported
-    test("exported", () {
+    test('exported', () {
       <dynamic>[
         OpenDatabaseOptions,
         DatabaseFactory,
@@ -13,6 +13,11 @@ void main() {
         Batch,
         ConflictAlgorithm,
         inMemoryDatabasePath,
+        OnDatabaseConfigureFn,
+        OnDatabaseCreateFn,
+        OnDatabaseOpenFn,
+        OnDatabaseVersionChangeFn,
+        onDatabaseDowngradeDelete,
       ].forEach((dynamic value) {
         expect(value, isNotNull);
       });